🎉 非常好的opc采集器，超越telegraf

+++ models/config/tag.go
@@ -3,7 +3,7 @@ package config
 import "time"
 
 type Tag struct {
-    Quality    int8      `json:"quality" xorm:"-"`
+    Quality    int      `json:"quality" xorm:"-"`
     Id         int64     `json:"id" xorm:"autoincr pk INTEGER"`
     Value      float64   `json:"value" xorm:"-"`
     Time       int64     `json:"time" xorm:"-"`
